adds discoverService
